Added function for tournaments; fixed bug
Added function to simulate tournaments. Also fixed a bug which sometimes caused MatchAllPlayers to have player 1 play player 1 when there were only 2 players. (The bug occurred because MatchAllPlayers was sometimes calling sample(2,1,replace=FALSE), which sampled not from the vector c(2) but rather from the vector 1:2 as the x argument was of length one.)